Move find_center... and do docstring cleanup.
